requirements.txt
langchain>=0.1.16
langchain-community>=0.0.21
openai>=1.0.0
python-dotenv